ID Card can be generated by using gdrive ZIP and excel sheet
The user needs to download full zip of the responses which contains photos and
signature folder. It needs to be saved here in the working directory and same
needs to be done for the excel sheet. Save both the files here and only one
file of both at a time.